requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.28,<1.0
psycopg[binary]>=3.1,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
boto3>=1.34,<2.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
gunicorn>=22.0,<23.0